Add memory package
